+++ package.json
@@ -96,8 +96,7 @@
     "tslib": "~2.3.1",
     "tsyringe": "^4.6.0",
     "tty-table": "^4.1.5",
-    "uuid": "^8.3.2",
-    "ws": "^8.17.1"
+    "uuid": "^8.3.2"
   },
   "devDependencies": {
     "@commitlint/cli": "^17.0.3",
@@ -119,7 +118,7 @@
     "@types/request-promise": "^4.1.48",
     "@types/semver": "^7.3.9",
     "@types/uuid": "^8.3.4",
-    "@types/ws": "^8.5.3",
+    "@types/ws": "^8.5.10",
     "@typescript-eslint/eslint-plugin": "5.33.1",
     "@typescript-eslint/parser": "5.33.1",
     "eslint": "8.15.0",
