+++ packages/repeater/src/api/RepeatersManager.ts
@@ -1,4 +1,6 @@
 export interface RepeatersManager {
+  getRepeater(repeaterId: string): Promise<{ repeaterId: string }>;
+
   createRepeater(options: {
     name: string;
     projectId?: string;
