+++ packages/repeater/src/api/commands/GetRepeaterRequest.ts
@@ -1,0 +1,20 @@
+import { HttpRequest } from '@sectester/bus';
+
+export interface GetRepeaterResponsePayload {
+  id: string;
+  name: string;
+  projectIds: string[];
+}
+
+export class GetRepeaterRequest extends HttpRequest<
+  undefined,
+  GetRepeaterResponsePayload
+> {
+  constructor(repeaterId: string) {
+    super({
+      url: `/api/v1/repeaters/${repeaterId}`,
+      method: 'GET',
+      payload: undefined
+    });
+  }
+}
