+++ packages/repeater/src/api/commands/index.ts
@@ -1,2 +1,3 @@
 export { CreateRepeaterRequest } from './CreateRepeaterRequest';
 export { DeleteRepeaterRequest } from './DeleteRepeaterRequest';
+export { GetRepeaterRequest } from './GetRepeaterRequest';
